topic updater enhanced test. evolution 67.sql

+++ kifi-backend/modules/common/conf/evolutions/shoebox/67.sql
@@ -4,8 +4,8 @@ CREATE TABLE topic_seq_num_info (
     id bigint(20) NOT NULL AUTO_INCREMENT,    
     created_at datetime NOT NULL,
     updated_at datetime NOT NULL,
-    uri_seq INT NOT NULL,
-    bookmark_seq INT NOT NULL,
+    uri_seq bigint(20) NOT NULL,
+    bookmark_seq bigint(20) NOT NULL,
     PRIMARY KEY (id)
 );
